requirements.txt
#annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
boto3==1.40.15
botocore==1.40.15
certifi==2025.8.3
distro==1.9.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
exceptiongroup==1.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
openai==1.101.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
